+++ Apartment/Class/DataModel/ZCPCoupletModel.h
@@ -23,14 +23,13 @@ typedef NS_ENUM(NSInteger, ZCPCoupletCollectState) {
 
 @interface ZCPCoupletModel : ZCPDataModel
 
-@property (nonatomic, assign) int coupletId;                        // 对联表编号
+@property (nonatomic, assign) NSInteger coupletId;                  // 对联表编号
 @property (nonatomic, copy) NSString *coupletContent;               // 对联内容
-@property (nonatomic, assign) int coupletReplyNumber;               // 回复人数
-@property (nonatomic, assign) int coupletCollectNumber;             // 收藏对联人数
-@property (nonatomic, assign) int coupletSupport;                   // 点赞量
+@property (nonatomic, assign) NSInteger coupletReplyNumber;         // 回复人数
+@property (nonatomic, assign) NSInteger coupletCollectNumber;       // 收藏对联人数
+@property (nonatomic, assign) NSInteger coupletSupport;             // 点赞量
 @property (nonatomic, strong) NSDate *coupletTime;                  // 出对时间
 @property (nonatomic, strong) ZCPUserModel *user;                   // 出对人
-
 @property (nonatomic, assign) ZCPCoupletSupportState supported;     // 当前用户是否已经点过赞
 @property (nonatomic, assign) ZCPCoupletCollectState collected;     // 当前用户是否已经收藏该对联
