+++ Apartment/Class/DataModel/ZCPStateModel.h
@@ -10,10 +10,10 @@
 
 @interface ZCPStateModel : ZCPDataModel
 
-@property (nonatomic, assign) int stateId;        // 状态表编号
-@property (nonatomic, copy) NSString *stateName;  // 状态名称
-@property (nonatomic, assign) int stateValue;     // 状态值
-@property (nonatomic, copy) NSString *stateType;  // 状态类型（标识是属于哪个模型的状态）
-@property (nonatomic, strong) NSDate *stateTime;  // 记录添加时间
+@property (nonatomic, assign) NSInteger stateId;        // 状态表编号
+@property (nonatomic, copy) NSString *stateName;        // 状态名称
+@property (nonatomic, assign) NSInteger stateValue;     // 状态值
+@property (nonatomic, copy) NSString *stateType;        // 状态类型（标识是属于哪个模型的状态）
+@property (nonatomic, strong) NSDate *stateTime;        // 记录添加时间
 
 @end
